@@ -1,6 +1,6 @@
 {
   "name": "@raggesilver/reqparams",
-  "version": "3.3.3",
+  "version": "4.0.0",
   "description": "Middleware to validate required parameters (POST/GET)",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -10,7 +10,8 @@
   "scripts": {
     "prepublishOnly": "npm run build",
     "build": "tsc",
-    "test": "jest"
+    "test": "jest --verbose",
+    "lint": "eslint"
   },
   "repository": "git@gitlab.com:raggesilver-npm/reqparams.git",
   "author": "Paulo Queiroz <pvaqueiroz@gmail.com>",
@@ -21,16 +22,22 @@
   },
   "homepage": "https://gitlab.com/raggesilver-npm/reqparams#README",
   "devDependencies": {
-    "@types/express": "^4.17.6",
-    "@types/jest": "^26.0.15",
-    "@types/mongoose": "^5.7.37",
-    "@types/node": "^14.6.0",
-    "axios": "^0.19.2",
+    "@types/express": "^4.17.11",
+    "@types/jest": "^26.0.20",
+    "@types/mongoose": "^5.10.3",
+    "@types/node": "^14.14.31",
+    "@types/supertest": "^2.0.10",
+    "@typescript-eslint/eslint-plugin": "^4.15.1",
+    "@typescript-eslint/parser": "^4.15.1",
+    "axios": "^0.21.1",
     "body-parser": "^1.19.0",
+    "eslint": "^7.20.0",
+    "eslint-plugin-jest": "^24.1.5",
     "express": "^4.17.1",
-    "jest": "^26.4.1",
-    "ts-jest": "^26.4.4",
-    "typescript": "^4.0.2"
+    "jest": "^26.6.3",
+    "supertest": "^6.1.3",
+    "ts-jest": "^26.5.1",
+    "typescript": "^4.1.5"
   },
   "dependencies": {
     "@raggesilver/hidash": "^2.0.2"
