@@ -10,9 +10,10 @@
 
 import * as express from 'express';
 import * as bodyParser from 'body-parser';
+
 import { randomBytes } from 'crypto';
 
-import { reqparams, reqall, reqquery, validId, notEmpty } from '../src';
+import { reqall, ParamBuilder } from '../src';
 
 const app = express();
 const port = Number(process.env.PORT) || 9143;
@@ -20,15 +21,15 @@ const port = Number(process.env.PORT) || 9143;
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 
-const postAMid = reqparams({
-  username: { validate: notEmpty },
+const postAMid = reqall('body', {
+  username: ParamBuilder.String().notEmpty(),
 });
 
 app.post('/a', postAMid, (_req, res) => {
   return res.json({});
 });
 
-const postNest = reqparams({
+const postNest = reqall('body', {
   'name.first': {},
   'name.last': {},
 });
@@ -38,7 +39,7 @@ app.post('/nest', postNest, (_req, res) => {
 });
 
 const strictMid = reqall('body', {
-  'some.key': { type: String },
+  'some.key': ParamBuilder.String(),
 }, {
   strict: true,
   onerror: (_req, res, _next, message) => {
@@ -51,25 +52,18 @@ const strictMid = reqall('body', {
 
 app.post('/strict', strictMid, (req, res) => res.json(req.body));
 
-const passReqMid = reqparams({
-  user: {
-    type: String,
-    validate: [
-      notEmpty,
-      (val, req) => {
-        req.body.user = `${val} -- I got more stuff`;
-        return (true);
-      },
-    ],
-  },
+const passReqMid = reqall('body', {
+  'user': ParamBuilder.String().notEmpty().setValidate((v, req) => {
+    req.body.user = `${v} -- I got more stuff`;
+    return true;
+  }),
 });
 
 app.post('/passreq', passReqMid, (req, res) => res.status(200).json(req.body));
 
-const resourceByIdMid = reqall(
-  'params',
-  { id: { type: String, validate: validId }}
-);
+const resourceByIdMid = reqall('params', {
+  id: ParamBuilder.ObjectId(),
+});
 
 app.get('/resource_by_id/:id', resourceByIdMid, (req, res) => {
   return res.status(200).json({
@@ -78,9 +72,9 @@ app.get('/resource_by_id/:id', resourceByIdMid, (req, res) => {
   });
 });
 
-const allowNull = reqparams({
-  name: { type: String, required: false }, // May be null
-  age: { type: Number, required: false, nullable: false, validate: v => v >= 18 },
+const allowNull = reqall('body', {
+  name: ParamBuilder.String().notRequired().setNullable(true), // May be null
+  age: ParamBuilder.Number({ integer: true }).notRequired().min(18),
 }, {
   strict: true,
 });
@@ -89,8 +83,8 @@ app.post('/allow_null', allowNull, (req, res) => {
   return res.json(req.body);
 });
 
-const requiredAndNullable = reqparams({
-  name: { type: String, nullable: true }, // May be null
+const requiredAndNullable = reqall('body', {
+  name: ParamBuilder.String().setNullable(true), // May be null
 }, {
   strict: true,
 });
@@ -99,59 +93,59 @@ app.post('/required_and_nullable', requiredAndNullable, (req, res) => {
   return res.json(req.body);
 });
 
-const requiredIf = reqparams({
-  'name.first': { requiredIf: { $exists: 'name' }},
-  'name.last': { requiredIf: { $exists: 'name' }},
+const requiredIf = reqall('body', {
+  'name.first': ParamBuilder.String().setRequiredIfExists('name'),
+  'name.last': ParamBuilder.String().setRequiredIfExists('name'),
 });
 
 app.post('/required_if', requiredIf, (req, res) => {
   return res.json(req.body);
 });
 
-const requiredIf2 = reqparams({
-  'user.name.first': { requiredIf: { $exists: 'user.name' }},
-  'user.name.last': { requiredIf: { $exists: 'user.name' }},
-  'user.address.line1': { requiredIf: { $exists: 'user' }},
+const requiredIf2 = reqall('body', {
+  'user.name.first': ParamBuilder.String().setRequiredIfExists('user.name'),
+  'user.name.last': ParamBuilder.String().setRequiredIfExists('user.name'),
+  'user.address.line1': ParamBuilder.String().setRequiredIfExists('user'),
 });
 
 app.post('/required_if2', requiredIf2, (req, res) => {
   return res.json(req.body);
 });
 
-const requiredIf3 = reqparams({
-  'user.name.first': { type: String },
-  'user.name.last': { type: String },
-  'user.partner.name.first': { type: String, validate: notEmpty, requiredIf: { $exists: 'user.partner' }},
-  'user.partner.name.last': { type: String, validate: notEmpty, requiredIf: { $exists: 'user.partner' }},
+const requiredIf3 = reqall('body', {
+  'user.name.first': ParamBuilder.String(),
+  'user.name.last': ParamBuilder.String(),
+  'user.partner.name.first': ParamBuilder.String().notEmpty().setRequiredIfExists('user.partner'),
+  'user.partner.name.last': ParamBuilder.String().notEmpty().setRequiredIfExists('user.partner'),
 });
 
 app.post('/required_if3', requiredIf3, (req, res) => {
   return res.json(req.body);
 });
 
-const eitherMid = reqparams({
-  email: { type: String, either: 1 },
-  phone: { type: String, either: 1},
-  password: { type: String },
+const eitherMid = reqall('body', {
+  email: ParamBuilder.String().setEither(1),
+  phone: ParamBuilder.String().setEither(1),
+  password: ParamBuilder.String(),
 });
 
 app.post('/either', eitherMid, (req, res) => {
   return res.json(req.body);
 });
 
-const newTodoMid = reqparams({
-  todo: { type: Array, validate: notEmpty },
-  name: { type: String, validate: notEmpty },
-  something: { required: false, validate: (v, r) => notEmpty(v, r) || 'Something wrong is not right', },
+const newTodoMid = reqall('body', {
+  todo: ParamBuilder.Array().notEmpty(),
+  name: ParamBuilder.String().notEmpty(),
+  something: ParamBuilder.Object().notRequired().notEmpty(),
 });
 
 app.post('/todo', newTodoMid, (req, res) => {
   return res.json(req.body);
 });
 
-const findTodoMid = reqquery({
-  author: { type: String, either: 1 },
-  date: { type: Date, either: 1 },
+const findTodoMid = reqall('query', {
+  author: ParamBuilder.String().setEither(1),
+  date: ParamBuilder.Date().setEither(1),
 });
 
 app.get('/todo/search', findTodoMid, (req, res) => {
@@ -159,21 +153,45 @@ app.get('/todo/search', findTodoMid, (req, res) => {
 });
 
 const notUnderageMid = reqall('body', {
-  dob: {
-    type: Date,
-    validate: (v: Date) => {
-      return (
-        (Date.now() - v.getTime() >= 1000 * 60 * 60 * 24 * 365 * 18)
-          ? true : 'Grow up first, kid'
-      );
-    },
-  },
+  dob: ParamBuilder.Date().max(Date.now() - (1000 * 60 * 60 * 24 * 365 * 18), 'Grow up first, kid'),
 });
 
 app.post('/not_underage', notUnderageMid, (req, res) => {
-  if (!(req.body.dob instanceof Date)) throw new Error('Dob is not a Date');
+  if (!(req.body.dob instanceof Date)) {
+    throw new Error('Dob is not a Date');
+  }
   return res.json(req.body);
 });
+
+const aWeekDay = reqall('query', {
+  day: ParamBuilder.String().setEnum(
+    ['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday']
+  ),
+});
+
+app.get('/my_daily_tasks', aWeekDay, (req, res) => {
+  return res.status(200).json(req.query);
+});
+
+const aWeekDay2 = reqall('query', {
+  day: ParamBuilder.String()
+    .setEnum(['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday'])
+    .setEnumCmp((a, b) => (
+      typeof a === 'string' && typeof b === 'string'
+      && a.toLowerCase() === b.toLowerCase()
+    )),
+});
+
+app.get('/my_daily_tasks2', aWeekDay2, (req, res) => {
+  return res.status(200).json(req.query);
+});
+
+const minMaxMid = reqall('body', {
+  age: ParamBuilder.Number({ integer: true }).min(18).max(24),
+  username: ParamBuilder.String().min(8).max(12),
+});
+
+app.post('/min-max', minMaxMid, (_, res) => res.status(200).json({}));
 
 const handler = app.listen(port);
 
@@ -186,22 +204,22 @@ const handler = app.listen(port);
 //        ╭━╯┃
 //  ----- ╰━━╯ _________________________________________________________________
 
-const axios = require('axios').default;
+import axios, { AxiosError } from 'axios';
 
 axios.defaults.baseURL = `http://127.0.0.1:${port}`;
 
 test('POST /a { username: notEmpty } FAIL', async () => {
   try {
-    await axios.post(`/a`);
+    await axios.post('/a');
   }
   catch (e) {
     expect(e.response.status).toBe(400);
-    expect(e.response.data).toMatchObject({ error: 'Parameter username missing' });
+    expect(e.response.data).toMatchObject({ error: 'username is required' });
   }
 });
 
 test('POST /a { username: notEmpty } OK', async () => {
-  let { data } = await axios.post(`/a`, {
+  const { data } = await axios.post('/a', {
     username: 'aa',
   });
   expect(data).toMatchObject({});
@@ -209,7 +227,7 @@ test('POST /a { username: notEmpty } OK', async () => {
 
 test('POST /nest { username: notEmpty } FAIL', async () => {
   try {
-    await axios.post(`/nest`, {
+    await axios.post('/nest', {
       name: {
         first: 'Paulo',
       },
@@ -218,13 +236,13 @@ test('POST /nest { username: notEmpty } FAIL', async () => {
   catch (e) {
     expect(e.response.status).toBe(400);
     expect(e.response.data).toMatchObject({
-      error: 'Parameter name.last missing',
+      error: 'name.last is required',
     });
   }
 });
 
 test('POST /nest { username: notEmpty } OK', async () => {
-  let { data } = await axios.post(`/nest`, {
+  const { data } = await axios.post('/nest', {
     name: {
       first: 'Paulo',
       last: 'Queiroz',
@@ -234,7 +252,7 @@ test('POST /nest { username: notEmpty } OK', async () => {
 });
 
 test('POST /strict { some.key: String } OK', async () => {
-  let { data } = await axios.post(`/strict`, {
+  const { data } = await axios.post('/strict', {
     some: {
       key: 'some value',
       removeMe: true,
@@ -255,7 +273,7 @@ test('POST /strict { some.key: String } FAIL', async () => {
   let error = undefined;
 
   try {
-    let { data: d } = await axios.post(`/strict`, {
+    const { data: d } = await axios.post('/strict', {
       some: {
         key: 13,
       },
@@ -275,7 +293,7 @@ test('POST /passreq { user: \'aaa\' }', async () => {
   expect.assertions(1);
   try {
     const user = 'aaa';
-    const { data } = await axios.post(`/passreq`, { user });
+    const { data } = await axios.post('/passreq', { user });
 
     expect(data?.user).toBe(`${user} -- I got more stuff`);
   }
@@ -293,7 +311,7 @@ test('GET /resource_by_id/:id FAIL', async () => {
     await axios.get(`/resource_by_id/${invalid_id}`);
   }
   catch (e) {
-    expect(e.response?.data.error).toBe('Invalid parameter id');
+    expect(e.response?.data.error).toBe('tttttttttttt is not a valid id');
   }
 });
 
@@ -304,7 +322,7 @@ test('GET /resource_by_id/:id FAIL 2', async () => {
     await axios.get(`/resource_by_id/${invalid_id}`);
   }
   catch (e) {
-    expect(e.response?.data.error).toBe('Invalid parameter id');
+    expect(e.response?.data.error).toBe('aaaaaaaaaaaaaaaaaaaaaaaaa is not a valid id');
   }
 });
 
@@ -329,7 +347,7 @@ test('POST /allow_null FAIL', async () => {
     await axios.post('/allow_null', { name: null, age: null });
   }
   catch (e) {
-    expect(e.response?.data.error).toBe('Invalid type for param \'age\'');
+    expect(e.response?.data.error).toBe('Invalid type for age');
   }
 });
 
@@ -372,7 +390,7 @@ test('POST /required_and_nullable FAIL', async () => {
     await axios.post('/required_and_nullable', {});
   }
   catch (e) {
-    expect(e.response?.data.error).toBe('Parameter name missing');
+    expect(e.response?.data.error).toBe('name is required');
   }
 });
 
@@ -395,7 +413,7 @@ test('POST /required_if FAIL', async () => {
     await axios.post('/required_if', { name: { whatever: 'Blah' }});
   }
   catch (e) {
-    expect(e.response?.data.error).toBe('name.first is required if name is present');
+    expect(e.response?.data.error).toBe('name.first is required when name is provided');
   }
 });
 
@@ -405,7 +423,7 @@ test('POST /required_if FAIL 2', async () => {
     await axios.post('/required_if', { name: { first: 'Hacker' }});
   }
   catch (e) {
-    expect(e.response?.data.error).toBe('name.last is required if name is present');
+    expect(e.response?.data.error).toBe('name.last is required when name is provided');
   }
 });
 
@@ -434,7 +452,7 @@ test('POST /required_if2 OK', async () => {
 test('POST /required_if3 FAIL', async () => {
   expect.assertions(1);
   try {
-    const res = await axios.post('/required_if3', {
+    await axios.post('/required_if3', {
       user: {
         name: {
           first: 'Mr',
@@ -447,7 +465,7 @@ test('POST /required_if3 FAIL', async () => {
     });
   }
   catch (e) {
-    expect(e.response?.data?.error).toBe('user.partner.name.first is required if user.partner is present');
+    expect(e.response?.data?.error).toBe('user.partner.name.first is required when user.partner is provided');
   }
 });
 
@@ -470,7 +488,7 @@ test('POST /either FAIL', async () => {
     await axios.post('/either', { password: 'sEcUrE123!' });
   }
   catch (e) {
-    expect(e.response?.data.error).toBe('At least one of email, phone must be present');
+    expect(e.response?.data.error).toBe('One of email or phone is required');
   }
 });
 
@@ -478,13 +496,8 @@ test('POST /either FAIL', async () => {
 
 test('POST /todo OK', async () => {
   expect.assertions(1);
-  try {
-    const res = await axios.post('/todo', { todo: ['Chores', 'Buy stuff', 'Homework'], name: 'Today\'s tasks' });
-    expect(res.status).toBe(200);
-  }
-  catch (e) {
-    console.error(e.response?.data?.error || e);
-  }
+  const res = await axios.post('/todo', { todo: ['Chores', 'Buy stuff', 'Homework'], name: 'Today\'s tasks' });
+  expect(res.status).toBe(200);
 });
 
 test('POST /todo FAIL', async () => {
@@ -493,17 +506,23 @@ test('POST /todo FAIL', async () => {
     await axios.post('/todo', { todo: [], name: 'I\'m lazy' });
   }
   catch (e) {
-    expect(e.response?.data.error).toBe('Invalid parameter todo');
+    expect(e.response?.data.error).toBe('todo may not be empty');
   }
 });
 
-test('POST /todo FAIL', async () => {
+test('POST /todo OK 2', async () => {
+  expect.assertions(1);
+  const res = await axios.post('/todo', { todo: ['AA'], name: 'I\'m not lazy', something: { 'willNotEmptyFail?': false }});
+  expect(res.status).toBe(200);
+});
+
+test('POST /todo FAIL 2', async () => {
   expect.assertions(1);
   try {
-    await axios.post('/todo', { todo: ['AA'], name: 'I\'m not lazy', something: { 'willNotEmptyFail?': true }});
+    await axios.post('/todo', { todo: ['AA'], name: 'I\'m not lazy', something: {}});
   }
   catch (e) {
-    expect(e.response?.data.error).toBe('Something wrong is not right');
+    expect(e.response?.data.error).toBe('something may not be empty');
   }
 });
 
@@ -518,7 +537,7 @@ test('GET /todo/search OK', async () => {
   }
 });
 
-test('GET /todo/search OK', async () => {
+test('GET /todo/search OK 2', async () => {
   expect.assertions(1);
   try {
     const res = await axios.get('/todo/search', { params: { date: new Date }});
@@ -532,16 +551,22 @@ test('GET /todo/search OK', async () => {
 // Test date values ============================================================
 
 test('POST /not_underage OK', async () => {
-  const dob = new Date();
+  expect.assertions(2);
+  try {
+    const dob = new Date();
 
-  dob.setFullYear(1990);
+    dob.setFullYear(1990);
 
-  const res = await axios.post('/not_underage', {
-    dob,
-  });
+    const res = await axios.post('/not_underage', {
+      dob,
+    });
 
-  expect(res.status).toBe(200);
-  expect(res.data?.dob).toBe(dob.toISOString());
+    expect(res.status).toBe(200);
+    expect(res.data?.dob).toBe(dob.toISOString());
+  }
+  catch (e) {
+    console.error(e.response.data);
+  }
 });
 
 test('POST /not_underage OK 2', async () => {
@@ -567,7 +592,7 @@ test('POST /not_underage FAIL', async () => {
   }
   catch (e) {
     expect(e.response?.status).toBe(400);
-    expect(e.response?.data.error).toBe('Invalid type for param \'dob\'');
+    expect(e.response?.data.error).toBe('Invalid type for dob');
   }
 });
 
@@ -580,7 +605,7 @@ test('POST /not_underage FAIL 2', async () => {
   }
   catch (e) {
     expect(e.response?.status).toBe(400);
-    expect(e.response?.data.error).toBe('Invalid type for param \'dob\'');
+    expect(e.response?.data.error).toBe('Invalid type for dob');
   }
 });
 
@@ -595,6 +620,117 @@ test('POST /not_underage FAIL underage', async () => {
     expect(e.response?.status).toBe(400);
     expect(e.response?.data.error).toBe('Grow up first, kid');
   }
+});
+
+test('GET /my_daily_tasks OK - testing enum', async () => {
+  const body = { day: 'Friday' };
+  const res = await axios.get('/my_daily_tasks', {
+    params: body,
+  });
+
+  expect(res.status).toBe(200);
+  expect(JSON.stringify(res.data)).toBe(JSON.stringify(body));
+});
+
+test('GET /my_daily_tasks FAIL - testing enum', async () => {
+  expect.assertions(2);
+  try {
+    await axios.get('/my_daily_tasks', {
+      params: { day: 'FRiday' },
+    });
+  }
+  catch (e) {
+    expect(e.response?.status).toBe(400);
+    expect(e.response?.data.error).toBe('day must be one of Monday, Tuesday, Wednesday, Thursday or Friday');
+  }
+});
+
+test('GET /my_daily_tasks FAIL 2 - testing enum', async () => {
+  expect.assertions(2);
+  try {
+    await axios.get('/my_daily_tasks', {
+      params: { day: 'BANANA' },
+    });
+  }
+  catch (e) {
+    expect(e.response?.status).toBe(400);
+    expect(e.response?.data.error).toBe('day must be one of Monday, Tuesday, Wednesday, Thursday or Friday');
+  }
+});
+
+test('GET /my_daily_tasks2 OK - testing enum', async () => {
+  const body = { day: 'fRIDAY' };
+  const res = await axios.get('/my_daily_tasks2', {
+    params: body,
+  });
+
+  expect(res.status).toBe(200);
+  expect(JSON.stringify(res.data)).toBe(JSON.stringify(body));
+});
+
+test('GET /my_daily_tasks2 FAIL - testing enum', async () => {
+  expect.assertions(2);
+  try {
+    await axios.get('/my_daily_tasks2', {
+      params: { day: 'BANANA' },
+    });
+  }
+  catch (e) {
+    expect(e.response?.status).toBe(400);
+    expect(e.response?.data.error).toBe('day must be one of Monday, Tuesday, Wednesday, Thursday or Friday');
+  }
+});
+
+test('POST /min-max OK', async () => {
+  expect.assertions(2);
+
+  const res = await axios.post('/min-max', {
+    age: 18,
+    username: 'la potato',
+  });
+
+  expect(res.status).toBe(200);
+  expect(res.data).toEqual({});
+});
+
+test('POST /min-max FAIL', async () => {
+  expect.assertions(8);
+
+  await axios.post('/min-max', {
+    age: 17,
+    username: 'la potato',
+  })
+    .catch((e: AxiosError) => {
+      expect(e.response?.status).toBe(400);
+      expect(e.response?.data.error).toBe('age must be at least 18');
+    });
+
+  await axios.post('/min-max', {
+    age: 25,
+    username: 'la potato',
+  })
+    .catch((e: AxiosError) => {
+      expect(e.response?.status).toBe(400);
+      expect(e.response?.data.error).toBe('age must be at most 24');
+    });
+
+  await axios.post('/min-max', {
+    age: 24,
+    username: 'la po',
+  })
+    .catch((e: AxiosError) => {
+      expect(e.response?.status).toBe(400);
+      expect(e.response?.data.error).toBe('username must be at least 8 characters long');
+    });
+
+  await axios.post('/min-max', {
+    age: 24,
+    username: 'la potato land of the beets',
+  })
+    .catch((e: AxiosError) => {
+      expect(e.response?.status).toBe(400);
+      expect(e.response?.data.error).toBe('username must be at most 12 characters long');
+    });
 });
 
 // Close listener after testing
